+++ packages/tokens/CHANGELOG.md
@@ -1,4 +1,16 @@
-# @bridger-ds/tokens
+# @bridger-kr/tokens
+
+## 0.1.2
+
+### Patch Changes
+
+- [#19](https://github.com/bridger-kr/bridger-design-system/pull/19) [`86f7d42`](https://github.com/bridger-kr/bridger-design-system/commit/86f7d42eddf083d1f5b39f98ea270d35eeaa6509) Thanks [@happycastle114](https://github.com/happycastle114)! - Remove the external Google Fonts import from the token font stylesheet so apps can keep CSP font and style policies self-hosted.
+
+## 0.1.1
+
+### Patch Changes
+
+- [#11](https://github.com/bridger-kr/bridger-design-system/pull/11) [`0e19cb4`](https://github.com/bridger-kr/bridger-design-system/commit/0e19cb480a379a3ff79bd5fcb842041518cc9b2f) Thanks [@happycastle114](https://github.com/happycastle114)! - Fix the status warning token to the Bridger product orange `#EC5E1F` so alert and app mirrors no longer introduce a second orange.
 
 ## 0.1.0